+++ src/components/admin/ConcertEditor.svelte
@@ -42,6 +42,9 @@
     }
 
     async function deleteConcert() {
+        const areYouSure = prompt(`If you really want to delete this concert (${location}), type YES and select ok`);
+        if (areYouSure !== "YES") return;
+
         await deleteDoc(concertRef);
         dispatch("deleted", {
             ref: concertRef
